add pretection for common services

+++ uninstall.sh
@@ -5,6 +5,9 @@ source lib/status.sh
 source 00-setup.sh
 source 99-uninstall.sh
 
+# To protect CS, we have a flag to indicate that
+all="${1}"
+
 # Uninstall Humio
 uninstall-humio
 
@@ -15,7 +18,9 @@ uninstall-event-manager
 uninstall-aimanager
 
 # Uninstall Common Services
+if [[ "$all" == "all" ]]; then
 uninstall-common-services
+fi
 
 # Uninstall post stuff
 uninstall-aimanager-post-actions
@@ -27,4 +32,6 @@ progress-bar 2
 # Delete projects
 oc delete project $NAMESPACE_HUMIO
 oc delete project $NAMESPACE_CP4AIOPS
+if [[ "$all" == "all" ]]; then
 oc delete project $NAMESPACE_CS
+fi
